+++ jest-preprocessor.js
@@ -3,6 +3,7 @@
  *
  * @format
  * @noflow
+ * @emails oncall+internationalization
  */
 const babel = require('@babel/core');
 const createCacheKeyFunction = require('fbjs-scripts/jest/createCacheKeyFunction');
@@ -12,7 +13,17 @@ const cacheKeyPackages = [
   'babel-preset-fbjs',
   'babel-plugin-fbt',
   'babel-plugin-fbt-runtime',
-].map(name => path.join(path.dirname(require.resolve(name)), 'package.json'));
+].map(name =>
+  path.join(
+    path.dirname(
+      // Find the actual module root from the package.json file,
+      // otherwise, the result may be incorrect if a custom "main" path was set.
+      // See https://stackoverflow.com/a/49455609/104598
+      require.resolve(path.join(name, 'package.json')),
+    ),
+    'package.json',
+  ),
+);
 
 // This is basically fbjs-scripts/jest/preprocessor, but with the
 // ability to specify additional plugins
@@ -25,7 +36,7 @@ function createTransformer(opts /*: Object */ = {}) {
           require('babel-preset-fbjs'),
         ],
         plugins: opts.plugins || [],
-        filename: filename,
+        filename,
         retainLines: true,
       };