+++ packages/fb-tiger-hash/src/index.js
@@ -9,12 +9,12 @@
 
 'use strict';
 
-const Tiger = require('./Tiger');
 const hashPhrases = require('./hashPhrases');
+const Tiger = require('./Tiger');
 
 module.exports = {
   __esModule: true,
   default: Tiger,
-  Tiger: Tiger,
-  hashPhrases: hashPhrases,
+  Tiger,
+  hashPhrases,
 };